First Commit after class with following activities from today
This is the first proper commit after today's class activities. This repository will host the Main project and follow along activities I attempt during class. At the start this repository will contain the follow along activities I was attempting today, as well as a .gitignore file to ignore the library folder (which can become extremely bloated)

The main project will be in a new folder called "MainGAM404Folder".
This commit will now be published to GitHub.