Setting the foundations of development down
This is a commit that follows on from the initial commit. I have set up the "MainGAM404Folder" and have put in a test scene and test script for the moment. As the development hasn't started just yet, I intend to flesh out my ideas from here in a brief, then develop the other parts. Will continue working over the week.