Added Development folder of follow-along activities from Week 2, as well as Brief+User Manual Documentation and UML Diagrams.
This commit added a new Unity Development folder that was used to help me follow along with the concepts introduced in Week 2. Although I am very familiar with Unity, this was a helpful series of activities that served as a good refresher and taught me some new tricks and tips with aspects such as PlayerPrefs, Canvases and UI, and loading scenes into the game.

This folder will be committed because it might be a good reference point for developing the main project as well. Additional documents added include the GAM405.1 Brief and the UML Documentation diagrams. Hopefully this will suffice and give me a good roadmap for developing the project.