Just some simple changes made to the Unity Fundamentals follow along file in class
This commit just contains some following-along activities from today's class in the Unity Fundamentals folder. Not really much to remark on, to be honest.

Starting from today I intend to begin work on the project in the main GAM405 development folder, and will likely begin populating the folder with the necessary building blocks for the project's development.