Basic Character Models and Item Models Made - Sounds added and Materials Made
This commit contains the basic characters and items made solely using Unity's geometric objects to make the characters and item models. This includes the player character, the two enemies, the batteries (four of each type), the stopwatch and the goal post.

New scenes have been made so that I can experiment and ensure that code is working properly. Following this commit, my next step will be to wire up the main menu so that it loads up the game's main scene, but will also tackle building the main code of the game so that there is control over the player's movement in 2D (X and Y)