Progress towards the functionality of the game and development of the main game scene.
Focused on developing the level and messing around with some aspects of platforming.
Implemented basic collision events for when the player collides with the tossing B.A.D.D.I enemy that will launch the player up and away in an arc. Additionally created a winning screen and losing screen for the game.
Initialised the ProBuilder Package to create a spikes prefab. If the player collides with the spike, it will be game over.