Major commit of relatively completed game level with populated enemies,
This is a fairly big commit containing a much more populated game level. There are more enemies lying around the stage and there are several more hazards added to the stages. Code has also been adjusted, and a new script has been made and attached to the camera. This code will contain the User Interface that will keep track of the player's stats, from health, lives, movement speed and jump  height. However this segment of the code has yet to be added, so it's just the main text at the moment.

The buttons on the end game screens (Game win and Game over) have now been wired up to work properly.

Made new folders to store any animations or animation controllers that are used in the game. The scripts have also been moved to their appropriate folder.

The next steps I might tackle could be the proper creation of the batteries as ENUMS and how they interact with the player (They increase the player's stats upon colliding with the player as a trigger), followed by implementing audio into the game.