Followed along with activities from Week 4 lessons - gives me good ideas for implementing audio properly into the game.
This commit is largely contained in the Unity Fundamentals folder, and was developed alongside the lesson material from Week 4 on Audio, Events and Prefabs. Completed these activities with success, and will carry this on to the main project later in the week to add audio and BGM to the main project files.