Post-Class Commit with the addition of songs.
For this commit, there isn't much that has changed since before the class today. There were some adjustments in the Unity Fundamentals Project, but mainly in regards to getting a better understanding of Events.

Following this I added three potential music sound tracks that I might consider using for the game. These have been sourced from FreeMusicArchive and have been properly credited and attributed in the credits documentation in the project. Over the course of tomorrow I will work towards implementing the audio, working on implementing events into the game project, and properly implementing the batteries and how they will engage/interact with the player and the UI.