Added sounds, music and began populating the stage with batteries. Timer and work done on User Interface.
For this commit, I have been working on implementing the sounds and music into the level. The songs have been properly attributed on the credits .txt file, and other adjustments have been made to make the sounds appropriate for the game, such as lowering the volume of the blast-off sound effect, which was noted to be loud.

Following this I then began working on implementing the timer that will take the player to the game over screen when the time limit of 3:00 minutes is gone. After this I might try to display the player's stats to the User Interface.

I began working on making the batteries work, but I still feel like I've been procrastinating on that point. That said, the batteries now have a separate 'Batteries_Version' class, intended to be used for the varying enum values of the batteries. I will also work towards making it so that when the player touches these batteries, they get the values added to their stats.

Also made a few adjustments to some other scenes.