Final commit of the day
Adjusted some of the final scenes to have more polish and walls so the player doesn't fly off bounds.

Additionally adjusted the collisions so that the batteries destroy themselves upon colliding with the player. Will ask about making it so that this then adds the value to the player's stats.