Miscellaneous addition of particle effect jetpack, and attempts at trying to make an event system work.
This commit has been focused on trying to make the event system work for the game. This involves readjusting the code so that the player's code has redundancies or functions that were empty or repetitive have been adjusted. A small particle system has been added to the player's backpack.

Attempts made to work on the enumerated batteries and Event handling, but the process is still incomplete. Will tackle over the final week of the project.

Added code for the goalpost to handle a scene transition to the winning screen instead of the player.