Follow along activities from Week 5  - Getting to the final stretch...
This commit is mainly consisting of the follow-along activities from Week 5 - focusing on elements such as raycasting, mouse clicking and responding to the mouse input, lighting, etc.

Some of these features I don't think will be in the game, but they are good skills to have on hand.

Hoping to solve the batteries over the course of this final week.