Post-class commit (will continue working on it in the rest of the afternoon)
This commit is mainly focused on adding the corrections that Edward suggested I make so that the batteries work as correct enumerations and also to use a Switch-Case system so that a message is displayed when the player collides with a certain BatteryType that influences the appropriate stat.

Additionally have also made it so that the player makes a sound when they absorb/collide with the battery.

Will continue to work on this in the rest of the afternoon, with a focus on making an event system work properly so that the user interface is updated with the player's stats, tidying up any redundant code, and making corrections to existing code.