Getting close to the end
This commit has added the proper functionality for the battery interactions with the player. They will now add the appropriate stats to the player when the player collides with them (Red = Launch distance, Blue = move speed, Green = health and Yellow = jump height)

However these are not finely tuned and can result in the player flying off at super high speeds.

Adjusted the layouts of the game level to add a bit more variety and some slight challenge. Blocks have been placed towards the ceiling to dissuade the player from always floating in the air.