FINAL COMMIT - Essentially finished product and new Manual file added
This is the final commit for submitting the assignment for GAM405.1.

I have successfully been able to add the player's stats to be tracked by the game system, so that the stats appear to the player during gameplay.
Collisions have also been slightly modified so that they are hopefully a bit more accurate.
A spawnpoint the player will return to upon touching the spikes has been added to the game, and the appropriate code has been added to the player.

The Unity Fundamentals Scenes have also been adjusted in the experimentation of adding UI updates from the player, and I am satisfied with how it works.

Additional screenshots and a brief manual have been added to the commit.
After this commit, the assignment will be complete.